+++ ChangeLog.md
@@ -3,6 +3,7 @@ The content of this file will be list all the tasks has done. I will update it a
 
 ### v1.0.1 - October 2nd, 2012
 * Add function to load jQuery library if Google CND not reachable
+* Add functions to load pro functions, custom functions, custom css if it exists in uploads folder and child theme folder.
 * Add ChangeLog file (this file) to logging the changes.
 * Rewrite IE conditional HTML5
 * Fix fatal bug to load external style from child theme when it's available